+++ packages/sdk-coverage-tests/src/index.js
@@ -1,4 +1,6 @@
 const {makeEmitTracker} = require('./code-export')
 const {makeCoverageTests} = require('./tests')
+const TestSetup = require('../coverage-tests/util/TestSetup')
+const ApiAssertions = require('../coverage-tests/util/ApiAssertions')
 
-module.exports = {makeEmitTracker, makeCoverageTests}
+module.exports = {makeEmitTracker, makeCoverageTests, TestSetup, ApiAssertions}